Rename app files for the static and dynamic segment pages
The convention is to use the name of the directory instead of `app.tsx`
for the files where the React app is initialized. See for example
`assets/js/src/wizard/wizard.tsx` and `assets/js/src/homepage/homepage.tsx`.

[MAILPOET-5392]
