+++ mailpoet/assets/js/src/segments/dynamic/dynamic.tsx
