+++ mailpoet/assets/js/src/segments/static/static.tsx
